move cosine_similary/2 to AI.Util because sharing is caring
